+++ tensorlayer/models/transformer/beamsearchHelper/beam_search.py
@@ -39,11 +39,6 @@ class SequenceBeamSearchV2(v1.SequenceBeamSearch):
         finished_scores = finished_state[_StateKeys.FINISHED_SCORES]
         finished_flags = finished_state[_StateKeys.FINISHED_FLAGS]
 
-        # # Account for corner case where there are no finished sequences for a
-        # # particular batch item. In that case, return alive sequences for that batch
-        # # item.
-        # finished_seq = tf.where(tf.reduce_any(finished_flags, 1), finished_seq, alive_seq)
-        # finished_scores = tf.where(tf.reduce_any(finished_flags, 1), finished_scores, alive_log_probs)
         return finished_seq, finished_scores
 
 
@@ -51,29 +46,38 @@ def sequence_beam_search(
         symbols_to_logits_fn, initial_ids, initial_cache, vocab_size, beam_size, alpha, max_decode_length, eos_id
 ):
     """Search for sequence of subtoken ids with the largest probability.
-
-    Args:
-        symbols_to_logits_fn: A function that takes in ids, index, and cache as
-        arguments. The passed in arguments will have shape:
+    
+    Parameters
+    -----------
+    symbols_to_logits_fn : A function with ids, index, and cache as arguments. 
+        The passed in arguments will have shape:
             ids -> [batch_size * beam_size, index]
             index -> [] (scalar)
             cache -> nested dictionary of tensors [batch_size * beam_size, ...]
         The function must return logits and new cache.
             logits -> [batch * beam_size, vocab_size]
             new cache -> same shape/structure as inputted cache
-        initial_ids: Starting ids for each batch item.
-        int32 tensor with shape [batch_size]
-        initial_cache: dict containing starting decoder variables information
-        vocab_size: int size of tokens
-        beam_size: int number of beams
-        alpha: float defining the strength of length normalization
-        max_decode_length: maximum length to decoded sequence
-        eos_id: int id of eos token, used to determine when a sequence has finished
+    initial_ids : int with shape [batch_size]
+        Starting ids for each batch item.
+    initial_cache: dict 
+        contain starting decoder variables information
+    vocab_size: int 
+        size of tokens
+    beam_size: int 
+        number of beams
+    alpha: float 
+        strength of length normalization
+    max_decode_length: int
+        maximum length to decoded sequence
+    eos_id: int 
+        id of eos token, used to determine when a sequence has finished
+    
+    Returns
+    -------
+    Top decoded sequences [batch_size, beam_size, max_decode_length]
+    sequence scores [batch_size, beam_size]
+  """
 
-    Returns:
-        Top decoded sequences [batch_size, beam_size, max_decode_length]
-        sequence scores [batch_size, beam_size]
-    """
     batch_size = tf.shape(initial_ids)[0]
 
     sbs = SequenceBeamSearchV2(
@@ -85,11 +89,14 @@ def sequence_beam_search(
 def _expand_to_same_rank(tensor, target):
     """Expands a given tensor to target's rank to be broadcastable.
 
-    Args:
+    Parameters
+    -----------
+    
         tensor: input tensor to tile. Shape: [b, d1, ..., da]
         target: target tensor. Shape: [b, d1, ..., da, ..., dn]
 
-    Returns:
+     Returns:
+    -----------
         Tiled tensor of shape [b, d1, ..., da, 1, ..., 1] with same rank of target.
 
     Raises: